Move employee filters from employee app to hardware app
